adding baseline args

+++ src/main/groovy/de/schemmea/ma/BaselineNfTest.groovy
@@ -44,7 +44,7 @@ class BaselineNfTest {
     @After
     public void cleanUp() {
         Plugins.stop()
-        if(currentFile){
+        if(currentFile && ! TestExecutor.ARGS.keepTestFile){
             currentFile.delete()
         }
     }
